Unify interactor angle generation & remove pill angles
Refactored the angle computation logic to consolidate ellipsoid (pill)
and full-range interactor angle generation into a single
implementation in HPUIInteractorFullRangeAngles. Removed the pill
angle specific classes and editors, merging their unique parameters
and inspector controls into FullRange.  This unifies angle
computation, reduces code duplication, and makes inspector controls
more flexible and maintainable. All usages now rely on the generalized
ComputeAngles, which supports arbitrary ellipsoid shapes using the a,
b, c axes. To get a spherical ray angle, the a, b, c would be set to
the same values.

This breaks any HPUIInteractorPillAngles assets. To fix, open the
HPUIInteractorPillAngles asset in a text editor, replace the value of
MonoBehaviour.m_Script.guid from "c4fcef0234684194ea61b3b79818aa08" to
"0b22705a00b17914e80f71aa7226f296"
