+++ Editor/HPUIInteractorFullRangeAnglesEditor.cs
@@ -12,7 +12,9 @@ namespace ubco.ovilab.HPUI.Editor
         private bool generating = false;
         private float maxAngle = 90;
         private float angleStep = 5;
-        private float raySelectionThreshold = 0.015f;
+        private float axisA = 0.015f;
+        private float axisB = 0.015f;
+        private float axisC = 0.015f;
 
         protected void OnEnable()
         {
@@ -42,13 +44,16 @@ namespace ubco.ovilab.HPUI.Editor
                 }
                 maxAngle = Mathf.Round(EditorGUILayout.Slider("Max Angle", maxAngle, 0, 180));
                 angleStep = Mathf.Round(EditorGUILayout.Slider("Angle Step", angleStep, 0, 180));
-                raySelectionThreshold = EditorGUILayout.Slider("Ray selection threshold", raySelectionThreshold, 0, 0.05f);
+
+                axisA = EditorGUILayout.Slider(new GUIContent("Length along local X", "Measures along the thumb's length"), axisA, 0, 0.05f);
+                axisB = EditorGUILayout.Slider(new GUIContent("Length along local Y", "Measures across the thumb's lateral thickness"), axisB, 0, 0.05f);
+                axisC = EditorGUILayout.Slider(new GUIContent("Length along local Z", "Measures through the thumb's depth (axial thickness)"), axisC, 0, 0.05f);
 
                 if (GUILayout.Button("Generate new values"))
                 {
                     if (EditorUtility.DisplayDialog("Generate angles", message, "Generate", "Cancel"))
                     {
-                        t.angles = HPUIInteractorFullRangeAngles.ComputeAngles((int)maxAngle, (int)angleStep, raySelectionThreshold);
+                        t.angles = HPUIInteractorFullRangeAngles.ComputeAngles((int)maxAngle, (int)angleStep, axisA, axisB, axisC);
                         EditorUtility.SetDirty(t);
                         serializedObject.ApplyModifiedProperties();
                     }
