--- Editor/HPUIInteractorPillAnglesEditor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9fc3ec4cc1a3ca94f8e1216f97ba5ba3
