--- Runtime/Interaction/HPUIInteractorPillAngles.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c4fcef0234684194ea61b3b79818aa08
